requirements/common.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Pillow==8.3.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
isort==5.9.1
black
psycopg2-binary==2.9.1